requirements.txt
firebase-admin
python-firebase
SpeechRecognition
PyAudio
requests
python-dotenv
Pillow